Made some improvements.
I added tostring methods to each registration model classes so that they can be easily found in a search. This also improves the look and feel when browsing the admin site.
I added a basic fetch api for the calendar so that we can block out dates that already taken. However, it still needs some work as I have some questions for Michael.