Added some more review/registration python code.
The next step is to override the admin form for stays so that we can customize the approval process. The goal is to remove the is_approved checkbox and instead add a button that sends a post request to the server. This post request will then set the is_approved flag while also creating the initial review object for that stay. At this point we would schedule the email to be sent out on completion of their stay. Once this is done, we will be able to use the review key to access the site and test the review page.